apps/agents/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Agent pipeline
anthropic==0.40.0
praw==7.8.1
tavily-python==0.5.0
pytrends==4.9.2
beautifulsoup4==4.12.3

# PDF generation
reportlab==4.2.5

# Cloud storage
cloudinary==1.41.0

# HTTP client
httpx==0.28.1

# Utilities
asyncio-throttle==1.0.2

# LangGraph orchestrator
langgraph>=0.2.0

# Web search (researcher / economist nodes)
ddgs>=1.0.0

# PPTX pitch deck generation
python-pptx>=1.0.0

# HTTP (used by scout_node alongside httpx)
requests>=2.32.0
